+++ src/components/destination/DestinationCard.vue
@@ -1,6 +1,6 @@
 <template>
   <div
-    v-if="textExists"
+    v-if="info"
     @click="selectDestination"
     class="card"
     style="width: 100%"
@@ -15,7 +15,7 @@
       <!-- textExists is {{ textExists }} -->
       {{ details }}
       <h5 class="card-title">Name:{{ details.properties.name }}</h5>
-      <p class="card-text">Details is {{ text }}</p>
+      <p class="card-text">Details is {{ info }}</p>
       <button @click="selectDestination" class="btn btn-info">
         Go to Destination
       </button>
@@ -25,50 +25,51 @@
 <script>
 import axios from "axios";
 import { ref } from "vue";
+import getDestinationInfo from "../../composables/destination/getDestinationInfo";
 
 export default {
   props: { details: Object },
   emits: ["selected"],
   setup(props, context) {
-    const textExists = ref(false);
-    const text = ref("");
+    // const loadText = async (placeId) => {
+    //   const options2 = {
+    //     method: "GET",
+    //     url: `https://opentripmap-places-v1.p.rapidapi.com/en/places/xid/${placeId}`,
+    //     headers: {
+    //       "X-RapidAPI-Key":
+    //         "480d26927amsh4735a5991d23945p1077e5jsnd9d743e8977c",
+    //       "X-RapidAPI-Host": "opentripmap-places-v1.p.rapidapi.com",
+    //     },
+    //   };
 
-    const loadText = async (placeId) => {
-      const options2 = {
-        method: "GET",
-        url: `https://opentripmap-places-v1.p.rapidapi.com/en/places/xid/${placeId}`,
-        headers: {
-          "X-RapidAPI-Key":
-            "480d26927amsh4735a5991d23945p1077e5jsnd9d743e8977c",
-          "X-RapidAPI-Host": "opentripmap-places-v1.p.rapidapi.com",
-        },
-      };
-
-      await axios
-        .request(options2)
-        .then(function (response) {
-          // console.log(response.data.image);
-          // var image = response.data.image;
-          console.log(
-            "Check if wikipedia exists",
-            response.data.wikipedia_extracts != "undefined"
-          );
-          if (response.data.wikipedia_extracts != "undefined") {
-            textExists.value = true;
-            text.value = response.data.wikipedia_extracts.text;
-          }
-          // var display = document.getElementById("display");
-          // display.setAttribute("src", image);
-          // var info = document.getElementById("info");
-          // info.innerHTML += text;
-        })
-        .catch(function (error) {
-          console.error(error);
-          textExists.value = false;
-        });
-    };
-
+    //   await axios
+    //     .request(options2)
+    //     .then(function (response) {
+    //       // console.log(response.data.image);
+    //       // var image = response.data.image;
+    //       console.log(
+    //         "Check if wikipedia exists",
+    //         response.data.wikipedia_extracts != "undefined"
+    //       );
+    //       if (response.data.wikipedia_extracts != "undefined") {
+    //         textExists.value = true;
+    //         text.value = response.data.wikipedia_extracts.text;
+    //       }
+    //       // var display = document.getElementById("display");
+    //       // display.setAttribute("src", image);
+    //       // var info = document.getElementById("info");
+    //       // info.innerHTML += text;
+    //     })
+    //     .catch(function (error) {
+    //       console.error(error);
+    //       textExists.value = false;
+    //     });
+    // };
+    const { loadText, error, info } = getDestinationInfo();
     loadText(props.details.properties.xid);
+
+    // loadText(props.details.properties.xid);
+
     const isClicked = ref(false);
 
     const selectDestination = () => {
@@ -79,7 +80,7 @@ export default {
       });
     };
 
-    return { selectDestination, text, textExists, isClicked };
+    return { selectDestination, info, isClicked };
   },
 };
 </script>
